requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
requests
beautifulsoup4
lxml
pandas
matplotlib
fpdf2